+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 - Bump stdlib to 0.5.2
 - `moved()` directives are now skipped if the target path already exists in the layout
 - `pcb publish` now uses single confirmation prompt instead of two
+- `pcb release` now works for boards without a layout directory
+
+### Removed
+
+- Remove `board_config.json` generation from `pcb release`
 
 ## [0.3.26] - 2026-01-20